AM-423_Return_role_acls_for_casetypes
