+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/internal/repositories/AccessManagementRepository.java
@@ -18,7 +18,6 @@ import uk.gov.hmcts.reform.amlib.internal.models.RoleBasedAccessRecord;
 import uk.gov.hmcts.reform.amlib.internal.models.query.AttributeData;
 import uk.gov.hmcts.reform.amlib.internal.repositories.mappers.JsonPointerMapper;
 import uk.gov.hmcts.reform.amlib.internal.repositories.mappers.PermissionSetMapper;
-import uk.gov.hmcts.reform.amlib.models.DefaultRolePermissions;
 import uk.gov.hmcts.reform.amlib.models.ExplicitAccessMetadata;
 import uk.gov.hmcts.reform.amlib.models.ResourceDefinition;
 
@@ -98,10 +97,6 @@ public interface AccessManagementRepository {
         + "and permissions & 2 = 2 and attribute = '' order by resource_id")
     List<String> getUserCases(String userId);
 
-    @SqlQuery("select role_name as role, permissions from default_permissions_for_roles where resource_type = 'case' and resource_name = :caseTypeId "
-        + "and attribute = '' order by role_name")
-    @RegisterConstructorMapper(DefaultRolePermissions.class)
-    List<DefaultRolePermissions> getRolePermissionsForCaseType(String caseTypeId);
 
     @SqlUpdate("insert into access_management_audit (access_management_id, resource_id, accessor_id, permissions, accessor_type, service_name, resource_type, resource_name, attribute, relationship, calling_service_name, audit_timestamp, changed_by, action) "
         + "values (:access_management_id, :resourceId, :accessorId, :permissionsAsInt, cast(:accessorType as accessor_type), :serviceName, :resourceType, :resourceName, :attributeAsString, :relationship,"
